+++ node/clients/Organizations.ts
@@ -1,0 +1,58 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import type { InstanceOptions, IOContext } from '@vtex/api'
+import { AppGraphQLClient } from '@vtex/api'
+
+import { QUERIES } from '../resolvers/Routes/queries'
+import { getTokenToHeader } from './index'
+
+export class OrganizationsGraphQLClient extends AppGraphQLClient {
+  constructor(ctx: IOContext, options?: InstanceOptions) {
+    super('vtex.b2b-organizations-graphql@0.x', ctx, options)
+  }
+
+  public getAddresses = async (costCenterId: string): Promise<any> => {
+    return this.graphql.query(
+      {
+        extensions: {
+          persistedQuery: {
+            provider: 'vtex.b2b-organizations-graphql@0.x',
+            sender: 'vtex.b2b-checkout-settings@1.x',
+          },
+        },
+        query: QUERIES.getAddresses,
+        variables: {
+          id: costCenterId,
+        },
+      },
+      {
+        params: {
+          headers: getTokenToHeader(this.context),
+          locale: this.context.locale,
+        },
+      }
+    )
+  }
+
+  public getOrganization = async (organizationId: string): Promise<any> => {
+    return this.graphql.query(
+      {
+        extensions: {
+          persistedQuery: {
+            provider: 'vtex.b2b-organizations-graphql@0.x',
+            sender: 'vtex.b2b-checkout-settings@1.x',
+          },
+        },
+        query: QUERIES.getOrganizationDetails,
+        variables: {
+          id: organizationId,
+        },
+      },
+      {
+        params: {
+          headers: getTokenToHeader(this.context),
+          locale: this.context.locale,
+        },
+      }
+    )
+  }
+}